inline_mode.py & only_admin.py got some little changes

+++ app/handlers/only_admin.py
@@ -1,4 +1,4 @@
-from aiogram.types import Message
+from aiogram.types import Message, CallbackQuery
 from aiogram.enums import ParseMode
 from functools import wraps
 
@@ -19,19 +19,24 @@ async def is_admin(user_id: int, type: str,  db) -> bool:
         logger.exception(f"Ошибка в функции is_admin [only_admin.py] [{e}]")
 
 def only_admin_access(db, type):
-    try:
         def decorator(func):
             @wraps(func)
-            async def wrapper(message: Message, *args, **kwargs):
-                if not await is_admin(message.from_user.id, type, db):
-                    logger.info(f"[{message.from_user.id}] Попытка использовать команду / нажать на кнопку [{message.from_user.first_name}]")
-                    await message.delete()
-                    await message.answer("🚫 <b>У вас нет доступа к этой команде</b>", parse_mode=ParseMode.HTML)
-                    return
-                return await func(message, *args, **kwargs)
+            async def wrapper(message: Message | CallbackQuery, *args, **kwargs):
+                try:
+                    if not await is_admin(message.from_user.id, type, db):
+                        logger.info(f"[{message.from_user.id}] Попытка использовать команду ИЛИ нажать на кнопку")
+
+                        if hasattr(message, "delete"):
+                            await message.delete()
+
+                        await message.answer("🚫 <b>У вас нет доступа к этой команде</b>", parse_mode=ParseMode.HTML)
+                        return
+
+                    return await func(message, *args, **kwargs)
+
+                except Exception as e:
+                    logger.exception(f"Ошибка в функции only_admin_access [only_admin.py] [{e}]")
+                    await message.answer("⚠️ <b>Что-то пошло не так. Попробуйте позже.</b>", parse_mode=ParseMode.HTML)
+
             return wrapper
         return decorator
-    except Exception as e:
-        logger.exception(f"Ошибка в функции only_admin_access [only_admin.py] [{e}]")
-
-#
